seperated the cmd file into a dedicated package
