+++ yamlator/cmd/__init__.py
@@ -1,0 +1,12 @@
+from yamlator.cmd.core import main
+from yamlator.cmd.core import validate_yaml_data_from_file
+from yamlator.cmd.core import display_violations
+from yamlator.cmd.core import DisplayMethod
+
+
+__all__ = [
+    'main',
+    'validate_yaml_data_from_file',
+    'display_violations',
+    'DisplayMethod'
+]
